+++ asilib/__init__.py
@@ -43,6 +43,7 @@ from asilib.analysis.equal_area import equal_area
 
 # Imager implementation functions and classes.
 from asilib.imager import Imager
+from asilib.imagers import Imagers
 from asilib.conjunction import Conjunction
 from asilib.asi.themis import themis, themis_info, themis_skymap
 from asilib.asi.rego import rego, rego_info, rego_skymap